Minifying and/or uglifying
since we can run the thing via `go run trivia.go` anyway...
