+++ .gitignore
@@ -2,4 +2,3 @@ scala/.idea/
 scala/scala.iml
 scala/out/
 *.class
-go/trivia/trivia
