--- go/Makefile
@@ -1,3 +1,0 @@
-build:
-	go build -x
-	cd trivia && go build -x
