+++ go/trivia.go
@@ -1,7 +1,9 @@
-package trivia
+package main
 
 import (
 	"fmt"
+	"math/rand"
+	"time"
 )
 
 type Game struct {
@@ -206,4 +208,31 @@ func (me *Game) WrongAnswer() bool {
 	}
 
 	return true
+}
+
+func main() {
+	notAWinner := false
+
+	game := NewGame()
+
+	game.Add("Chet")
+	game.Add("Pat")
+	game.Add("Sue")
+
+	rand.Seed(time.Now().UTC().UnixNano())
+
+	for {
+		game.Roll(rand.Intn(5) + 1)
+
+		if rand.Intn(9) == 7 {
+			notAWinner = game.WrongAnswer()
+		} else {
+			notAWinner = game.WasCorrectlyAnswered()
+
+		}
+
+		if !notAWinner {
+			break
+		}
+	}
 }
